+++ components/Article/TextBody.js
@@ -7,19 +7,48 @@ import Image from '../Image';
 import Paragraph from './Paragraph';
 import {parseBody} from './Utils';
 
-const TextBody = ({content, embeds}) => {
+const TextBody = ({AdComponent, adsblocks, content, domain, embeds, lazy}) => {
   
   if (!content) return null;
-  
   const bodyItems = parseBody(content);
+  
+  let ads_p = 0;
+  let ads_t = adsblocks.length;
+  let count_p = 0;
+  const RenderAds = () => {
+    ads_p++;
+    if(ads_p > ads_t)
+      return false;
+    return <AdComponent content={adsblocks[ads_p - 1]} />;
+  };
+  const RenderParagraph = ({value}) => {
+    let has_ads = false;
+    count_p++;
+    if(count_p === 2) {
+      count_p = 0;
+      has_ads = true;
+    }
+    return (
+      <>
+        <Paragraph value={value} />
+        {has_ads && <RenderAds />}
+      </>
+    );
+  };
+
   return map(bodyItems, ({type, value}, key) => {
     switch(type) {
-      case 'Paragraph':
-        return <Paragraph key={key} value={value} />;
+      case 'Paragraph': 
+        return <RenderParagraph key={key} value={value} />;
       case 'Facebook':
         return embeds && embeds.Facebook && <embeds.Facebook key={key} value={value} />;
       case 'Image':
-        return <Block custom="article-image-embed"><Image key={key} value={value} /></Block>;
+      case 'ImageLegacy':
+        return (
+          <Block custom="article-image-embed" key={key} mb="3">
+            <Image custom="image-article" content={value} domain={domain} lazy={lazy} />
+          </Block>
+        );
       case 'Instagram':
         return embeds && embeds.Instagram && <embeds.Instagram key={key} value={value} />;
       case 'Tweet':
@@ -31,7 +60,10 @@ const TextBody = ({content, embeds}) => {
 };
 
 TextBody.propTypes = {
+  AdComponent: PropTypes.func,
+  adsblocks: PropTypes.array,
   content: PropTypes.string.isRequired,
+  domain: PropTypes.string,
   embeds: PropTypes.object
 };