+++ components/Menus/SideItems.js
@@ -2,6 +2,7 @@ import {map} from 'lodash';
 import PropTypes from 'prop-types';
 import React from 'react';
 
+import SvgIcArrowMenu from '../../icons/IcArrowMenu';
 import Block from '../Block';
 
 const SideMenuItems = ({content, functions}) => {
@@ -9,11 +10,12 @@ const SideMenuItems = ({content, functions}) => {
   const Item = ({item}) => {
     const hasSubitems = item.subitems.length > 0; 
     return (
-      <Block custom="sidemenu-group" mb="4">
-        <Block custom="sidemenu-group-item" pb="2" onClick={() => functions.onMenuItem(item)} >
+      <Block custom="sidemenu-group" mb="3">
+        <Block custom="sidemenu-group-item" pb="2" align='row between middle' onClick={() => functions.onMenuItem(item)} >
           {!hasSubitems ? 
             <a href={item.path}>{item.name}</a> :
             <span href={item.path}>{item.name}</span>}
+          <SvgIcArrowMenu />
         </Block>
       </Block>
     );