+++ components/SectionTitle/index.js
@@ -4,14 +4,15 @@ import React from 'react';
 import Block from '../Block';
 import Typography from '../Typography';
 
-const SectionTitle = ({name}) => {
+const SectionTitle = ({custom, name}) => {
   return (
-    <Block mb="2">
-      <Typography custom="section-title">{name}</Typography>
+    <Block custom="section-title-block" mb="3">
+      <Typography custom={`section-title ${custom}`}>{name}</Typography>
     </Block>
   );
 };
 SectionTitle.propTypes = {
+  custom: PropTypes.string,
   name: PropTypes.string
 };
 export default SectionTitle;