+++ components/Templates/GridNews.js
@@ -3,18 +3,23 @@ import PropTypes from 'prop-types';
 import React from 'react';
 
 import Block from '../Block';
-import {Teaser} from '../Teasers';
+import SectionTitle from '../SectionTitle';
+import Teaser from '../Teasers';
 
 const GridNews = props => {
-  const {content, status} = props;
-  const {items} = content;
+  const {content, domain, lazy, status} = props;
+  const {items, title} = content;
   
   const propsTemplate = {
     align: 'between', 
     custom: 'templates-newsgrid', 
-    md: {align: 'row'}
+    mb: '6',
+    md: {align: 'row', mb: '5'}
   };
   
+  if(items.length === 2) {
+    propsTemplate.custom = 'templates-newsgrid two';
+  }
   if(items.length === 3) {
     propsTemplate.custom = 'templates-newsgrid three';
   }
@@ -23,20 +28,27 @@ const GridNews = props => {
   }
 
   return (
-    <Block {...propsTemplate}>
-      {map(items, (item, key) =>
-        <Teaser 
-          key={key}
-          content={item}
-          hasImageTop={true}
-          status={status}
-        />
-      )}
-    </Block>
+    <>
+      {title && title !== ''&& <SectionTitle name={title} />}
+      <Block {...propsTemplate}>
+        {map(items, (item, key) =>
+          <Teaser 
+            key={key}
+            content={item}
+            domain={domain}
+            hasImageTop={true}
+            lazy={lazy}
+            status={status}
+          />
+        )}
+      </Block>
+    </>
   );
 };
 GridNews.propTypes = {
   content: PropTypes.object,
+  domain: PropTypes.string,
+  lazy: PropTypes.func,
   status: PropTypes.object
 };
 export default GridNews;