+++ components/Templates/Subjects.js
@@ -3,32 +3,44 @@ import PropTypes from 'prop-types';
 import React from 'react';
 
 import Block from '../Block';
-import {Teaser} from '../Teasers';
-
+import SectionTitle from '../SectionTitle';
+import Teaser from '../Teasers';
 
 const Subjects = props => {
-  const {content, status} = props;
+  const {content, domain, lazy, ReadMore, status} = props;
+  const {title} = content;
   const propsTemplate = {custom: 'templates-subjects', lg:{align: 'row between'}};
   return (
-    <Block {...propsTemplate} > 
-      <Block custom='col left'>
-        {map(content['items-left'], (item, key) =>
-          <Teaser content={item} key={key} status={status} />)}
+    <>
+      {title && title !== ''&& <SectionTitle name={title} />}
+      <Block {...propsTemplate} > 
+        <Block custom={`col left ${content['style-left']}`}>
+          <SectionTitle name={content['title-left']} />
+          {map(content['items-left'], (item, key) =>
+            <Teaser content={item} domain={domain} lazy={lazy} key={key} status={status} />)}
+          {ReadMore && <ReadMore />}
+        </Block>
+        <Block custom={`col center ${content['style-center']}`}>
+          <SectionTitle name={content['title-center']} />
+          {map(content['items-center'], (item, key) =>
+            <Teaser content={item} domain={domain} lazy={lazy} key={key} status={status} />)}
+          {ReadMore && <ReadMore />}
+        </Block>
+        <Block custom={`col right ${content['style-right']}`}>
+          <SectionTitle name={content['title-right']} />
+          {map(content['items-right'], (item, key) =>
+            <Teaser content={item} domain={domain} lazy={lazy} key={key} status={status} />)}
+          {ReadMore && <ReadMore />}
+        </Block>
       </Block>
-      <Block custom='col center'>
-        {map(content['items-center'], (item, key) =>
-          <Teaser content={item} key={key} status={status} />)}
-      </Block>
-      <Block custom='col right'>
-        {map(content['items-right'], (item, key) =>
-          <Teaser content={item} key={key} status={status} />)}
-      </Block>
-    </Block>
-
+    </>
   );
 };
 Subjects.propTypes = {
   content: PropTypes.object,
+  domain: PropTypes.string,
+  lazy: PropTypes.func,
+  ReadMore: PropTypes.func,
   status: PropTypes.object
 };
 export default Subjects;