+++ components/Templates/index.js
@@ -1,5 +1,7 @@
+export {default as Columnists} from './Columnists';
 export {default as Featured} from './Featured';
 export {default as GridNews} from './GridNews';
-export {default as NewsList} from './ListNews';
+export {default as Latest} from './Latest';
+export {default as MostRead} from './MostRead';
+export {default as Related} from './Related';
 export {default as Subjects} from './Subjects';
-export {default as Columnists} from './Columnists';