+++ stories/Image.stories.js
@@ -2,7 +2,8 @@ import {withKnobs} from '@storybook/addon-knobs';
 import React from 'react';
 
 import Image from '../components/Image';
-import ImageTeaser from '../components/Image/Teaser';
+import ImageBackground from '../components/Image/Background';
+import imageMockup from '../mockup/image.json';
 import {sassWrapper} from '../utils/SassWrapper';
 
 export default {
@@ -20,25 +21,41 @@ export const Readme = () => {
 };
 
 export const Default = () => {
-  const props = {
-    value: {
-      'image-byline': 'Carlos Delgado',
-      'image-contentId': '1.125.1595196595',
-      'image-subtitle': '',
-      'image-subtitle-original': 'Diego Costa and Angel Trujillo1' 
-    }
+  const propsAbsolute = {
+    content: imageMockup[0],
+    lazyLoad: null
   };
-  return <Image {...props} />;
+  const propsContentId = {
+    content: imageMockup[3],
+    lazyLoad: null
+  };
+  return (
+    <>
+      <pre>Image {imageMockup[0]['image-path']}</pre>
+      <Image {...propsAbsolute} />
+      <pre>Image {imageMockup[3]['image-contentId']}</pre>
+      <Image {...propsContentId} />
+    </>
+  );
 };
 
 export const Background = () => {
-  const props = {
-    value: {
-      'image-byline': 'Carlos Delgado',
-      'image-contentId': '1.125.1595196595',
-      'image-subtitle': '',
-      'image-subtitle-original': 'Diego Costa and Angel Trujillo1' 
-    }
+  const propsAbsolute = {
+    content: imageMockup[1],
+    height: 380,
+    lazyLoad: null
   };
-  return <ImageTeaser {...props} />;
+  const propsContentId = {
+    content: imageMockup[2],
+    height: 380,
+    lazyLoad: null
+  };
+  return (
+    <>
+      <pre>Image.background {imageMockup[1]['image-path']}</pre>
+      <ImageBackground {...propsAbsolute} />
+      <pre>Image.background {imageMockup[2]['image-contentId']}</pre>
+      <ImageBackground {...propsContentId} />
+    </>
+  );
 };